Revised version with simplified design and bug fixes. urlHash is added.
Head and bottom bars are simplified. Locate me is run only if requested and a marker is added to the map. The info pannel is now based on bootstrap modal window.

The Wifi hospots list is provided from the Overpass API service.  With the addition of th urlHash, it is now possible to refresh the application to request new hospots from the Overpass service.  This solution is preferred to openlayers bbox strategy to diminush the number of requests to the Overpass service.